use column_type instead of data_type for determining MySQL -> Haskell type mapping
This is meant to be a prototype showing how we can use column_type
instead of data_type to determine what type a particular column should
have. This allows mappings like tinyint(1) -> Bool without mapping
tinyint(3) to Bool, for example.

Probably this implementation is bad because it will break existing code,
so a more clean transition path should probably be investigated.